- Change all c++std flags to 17 - Change GateType to be an `enum` instead of `enum class` (removes type conversion warnings) - Move GateVTable into its own file - Delete GateVTable's default constructor so it cannot be accidentally uninitialized - Verify all gates are defined when initializing a vtable when compiling in debug mode - Make GateVTable's template argument define the exact function type to store, instead of indirectly specifying part of the function type - Extract the vtable data definitions into functions - Checked benchmarks are okay
