+++ BUILD
@@ -44,6 +44,9 @@ PYBIND_FILES = glob(
 cc_library(
     name = "stim_lib",
     srcs = SOURCE_FILES_NO_MAIN,
+    copts = [
+        "-std=c++17",
+    ],
     includes = ["src/"],
     linkopts = ["-lpthread"],
 )
@@ -52,6 +55,7 @@ cc_binary(
     name = "stim",
     srcs = SOURCE_FILES_NO_MAIN + glob(["src/**/main.cc"]),
     copts = [
+        "-std=c++17",
         "-march=native",
         "-O3",
     ],
@@ -63,6 +67,7 @@ cc_binary(
     name = "stim_benchmark",
     srcs = SOURCE_FILES_NO_MAIN + PERF_FILES,
     copts = [
+        "-std=c++17",
         "-march=native",
         "-O3",
     ],
@@ -74,6 +79,7 @@ cc_test(
     name = "stim_test",
     srcs = SOURCE_FILES_NO_MAIN + TEST_FILES,
     copts = [
+        "-std=c++17",
         "-march=native",
     ],
     data = glob(["testdata/**"]),
@@ -90,6 +96,7 @@ cc_binary(
     srcs = SOURCE_FILES_NO_MAIN + PYBIND_FILES,
     copts = [
         "-O3",
+        "-std=c++17",
         "-fvisibility=hidden",
         "-march=native",
         "-DSTIM_PYBIND11_MODULE_NAME=stim",
