+++ CMakeLists.txt
@@ -72,8 +72,8 @@ find_package(GTest QUIET)
 if(${GTest_FOUND})
     add_executable(stim_test ${SOURCE_FILES_NO_MAIN} ${TEST_FILES})
     target_link_libraries(stim_test GTest::gtest GTest::gtest_main)
-    target_compile_options(stim_test PRIVATE -Wall -Wpedantic -fno-strict-aliasing -fsanitize=undefined -fsanitize=address ${MACHINE_FLAG})
-    target_link_options(stim_test PRIVATE -fsanitize=undefined -fsanitize=address -pthread)
+    target_compile_options(stim_test PRIVATE -Wall -Wpedantic -g -fno-omit-frame-pointer -fno-strict-aliasing -fsanitize=undefined -fsanitize=address ${MACHINE_FLAG})
+    target_link_options(stim_test PRIVATE -g -fno-omit-frame-pointer -fsanitize=undefined -fsanitize=address -pthread)
 
     add_executable(stim_test_o3 ${SOURCE_FILES_NO_MAIN} ${TEST_FILES})
     target_link_libraries(stim_test_o3 GTest::gtest GTest::gtest_main)
