+++ glue/javascript/build_wasm.sh
@@ -29,6 +29,7 @@ echo '</script>' >> out/all_stim_tests.html
 
 # Build web assembly module using emscripten.
 emcc \
+    -std=c++17 \
     -s NO_DISABLE_EXCEPTION_CATCHING \
     -s EXPORT_NAME="load_stim_module"  \
     -s MODULARIZE=1  \
