+++ src/stim/simulators/error_analyzer.cc
@@ -22,6 +22,71 @@
 
 using namespace stim;
 
+constexpr GateVTable<void (ErrorAnalyzer::*)(const OperationData&)> error_analyzer_vtable_data() {
+    return {{{
+        {GateType::DETECTOR, &ErrorAnalyzer::DETECTOR},
+        {GateType::OBSERVABLE_INCLUDE, &ErrorAnalyzer::OBSERVABLE_INCLUDE},
+        {GateType::TICK, &ErrorAnalyzer::TICK},
+        {GateType::QUBIT_COORDS, &ErrorAnalyzer::I},
+        {GateType::SHIFT_COORDS, &ErrorAnalyzer::SHIFT_COORDS},
+        {GateType::REPEAT, &ErrorAnalyzer::I},
+        {GateType::MX, &ErrorAnalyzer::MX},
+        {GateType::MY, &ErrorAnalyzer::MY},
+        {GateType::M, &ErrorAnalyzer::MZ},
+        {GateType::MRX, &ErrorAnalyzer::MRX},
+        {GateType::MRY, &ErrorAnalyzer::MRY},
+        {GateType::MR, &ErrorAnalyzer::MRZ},
+        {GateType::RX, &ErrorAnalyzer::RX},
+        {GateType::RY, &ErrorAnalyzer::RY},
+        {GateType::R, &ErrorAnalyzer::RZ},
+        {GateType::MPP, &ErrorAnalyzer::MPP},
+        {GateType::XCX, &ErrorAnalyzer::XCX},
+        {GateType::XCY, &ErrorAnalyzer::XCY},
+        {GateType::XCZ, &ErrorAnalyzer::XCZ},
+        {GateType::YCX, &ErrorAnalyzer::YCX},
+        {GateType::YCY, &ErrorAnalyzer::YCY},
+        {GateType::YCZ, &ErrorAnalyzer::YCZ},
+        {GateType::CX, &ErrorAnalyzer::ZCX},
+        {GateType::CY, &ErrorAnalyzer::ZCY},
+        {GateType::CZ, &ErrorAnalyzer::ZCZ},
+        {GateType::H, &ErrorAnalyzer::H_XZ},
+        {GateType::H_XY, &ErrorAnalyzer::H_XY},
+        {GateType::H_YZ, &ErrorAnalyzer::H_YZ},
+        {GateType::DEPOLARIZE1, &ErrorAnalyzer::DEPOLARIZE1},
+        {GateType::DEPOLARIZE2, &ErrorAnalyzer::DEPOLARIZE2},
+        {GateType::X_ERROR, &ErrorAnalyzer::X_ERROR},
+        {GateType::Y_ERROR, &ErrorAnalyzer::Y_ERROR},
+        {GateType::Z_ERROR, &ErrorAnalyzer::Z_ERROR},
+        {GateType::PAULI_CHANNEL_1, &ErrorAnalyzer::PAULI_CHANNEL_1},
+        {GateType::PAULI_CHANNEL_2, &ErrorAnalyzer::PAULI_CHANNEL_2},
+        {GateType::E, &ErrorAnalyzer::CORRELATED_ERROR},
+        {GateType::ELSE_CORRELATED_ERROR, &ErrorAnalyzer::ELSE_CORRELATED_ERROR},
+        {GateType::I, &ErrorAnalyzer::I},
+        {GateType::X, &ErrorAnalyzer::I},
+        {GateType::Y, &ErrorAnalyzer::I},
+        {GateType::Z, &ErrorAnalyzer::I},
+        {GateType::C_XYZ, &ErrorAnalyzer::C_XYZ},
+        {GateType::C_ZYX, &ErrorAnalyzer::C_ZYX},
+        {GateType::SQRT_X, &ErrorAnalyzer::H_YZ},
+        {GateType::SQRT_X_DAG, &ErrorAnalyzer::H_YZ},
+        {GateType::SQRT_Y, &ErrorAnalyzer::H_XZ},
+        {GateType::SQRT_Y_DAG, &ErrorAnalyzer::H_XZ},
+        {GateType::S, &ErrorAnalyzer::H_XY},
+        {GateType::S_DAG, &ErrorAnalyzer::H_XY},
+        {GateType::SQRT_XX, &ErrorAnalyzer::SQRT_XX},
+        {GateType::SQRT_XX_DAG, &ErrorAnalyzer::SQRT_XX},
+        {GateType::SQRT_YY, &ErrorAnalyzer::SQRT_YY},
+        {GateType::SQRT_YY_DAG, &ErrorAnalyzer::SQRT_YY},
+        {GateType::SQRT_ZZ, &ErrorAnalyzer::SQRT_ZZ},
+        {GateType::SQRT_ZZ_DAG, &ErrorAnalyzer::SQRT_ZZ},
+        {GateType::SWAP, &ErrorAnalyzer::SWAP},
+        {GateType::ISWAP, &ErrorAnalyzer::ISWAP},
+        {GateType::ISWAP_DAG, &ErrorAnalyzer::ISWAP},
+        {GateType::CXSWAP, &ErrorAnalyzer::CXSWAP},
+        {GateType::SWAPCX, &ErrorAnalyzer::SWAPCX},
+    }}};
+}
+
 void ErrorAnalyzer::remove_gauge(ConstPointerRange<DemTarget> sorted) {
     if (sorted.empty()) {
         return;
@@ -389,68 +454,7 @@ ErrorAnalyzer::ErrorAnalyzer(
       ignore_decomposition_failures(ignore_decomposition_failures),
       block_decomposition_from_introducing_remnant_edges(block_decomposition_from_introducing_remnant_edges),
       num_ticks_in_past(num_ticks),
-      gate_vtable(GateVTable<ErrorAnalyzer>({
-          {GateType::DETECTOR, &ErrorAnalyzer::DETECTOR},
-          {GateType::OBSERVABLE_INCLUDE, &ErrorAnalyzer::OBSERVABLE_INCLUDE},
-          {GateType::TICK, &ErrorAnalyzer::TICK},
-          {GateType::QUBIT_COORDS, &ErrorAnalyzer::I},
-          {GateType::SHIFT_COORDS, &ErrorAnalyzer::SHIFT_COORDS},
-          {GateType::REPEAT, &ErrorAnalyzer::I},
-          {GateType::MX, &ErrorAnalyzer::MX},
-          {GateType::MY, &ErrorAnalyzer::MY},
-          {GateType::M, &ErrorAnalyzer::MZ},
-          {GateType::MRX, &ErrorAnalyzer::MRX},
-          {GateType::MRY, &ErrorAnalyzer::MRY},
-          {GateType::MR, &ErrorAnalyzer::MRZ},
-          {GateType::RX, &ErrorAnalyzer::RX},
-          {GateType::RY, &ErrorAnalyzer::RY},
-          {GateType::R, &ErrorAnalyzer::RZ},
-          {GateType::MPP, &ErrorAnalyzer::MPP},
-          {GateType::XCX, &ErrorAnalyzer::XCX},
-          {GateType::XCY, &ErrorAnalyzer::XCY},
-          {GateType::XCZ, &ErrorAnalyzer::XCZ},
-          {GateType::YCX, &ErrorAnalyzer::YCX},
-          {GateType::YCY, &ErrorAnalyzer::YCY},
-          {GateType::YCZ, &ErrorAnalyzer::YCZ},
-          {GateType::CX, &ErrorAnalyzer::ZCX},
-          {GateType::CY, &ErrorAnalyzer::ZCY},
-          {GateType::CZ, &ErrorAnalyzer::ZCZ},
-          {GateType::H, &ErrorAnalyzer::H_XZ},
-          {GateType::H_XY, &ErrorAnalyzer::H_XY},
-          {GateType::H_YZ, &ErrorAnalyzer::H_YZ},
-          {GateType::DEPOLARIZE1, &ErrorAnalyzer::DEPOLARIZE1},
-          {GateType::DEPOLARIZE2, &ErrorAnalyzer::DEPOLARIZE2},
-          {GateType::X_ERROR, &ErrorAnalyzer::X_ERROR},
-          {GateType::Y_ERROR, &ErrorAnalyzer::Y_ERROR},
-          {GateType::Z_ERROR, &ErrorAnalyzer::Z_ERROR},
-          {GateType::PAULI_CHANNEL_1, &ErrorAnalyzer::PAULI_CHANNEL_1},
-          {GateType::PAULI_CHANNEL_2, &ErrorAnalyzer::PAULI_CHANNEL_2},
-          {GateType::E, &ErrorAnalyzer::CORRELATED_ERROR},
-          {GateType::ELSE_CORRELATED_ERROR, &ErrorAnalyzer::ELSE_CORRELATED_ERROR},
-          {GateType::I, &ErrorAnalyzer::I},
-          {GateType::X, &ErrorAnalyzer::I},
-          {GateType::Y, &ErrorAnalyzer::I},
-          {GateType::Z, &ErrorAnalyzer::I},
-          {GateType::C_XYZ, &ErrorAnalyzer::C_XYZ},
-          {GateType::C_ZYX, &ErrorAnalyzer::C_ZYX},
-          {GateType::SQRT_X, &ErrorAnalyzer::H_YZ},
-          {GateType::SQRT_X_DAG, &ErrorAnalyzer::H_YZ},
-          {GateType::SQRT_Y, &ErrorAnalyzer::H_XZ},
-          {GateType::SQRT_Y_DAG, &ErrorAnalyzer::H_XZ},
-          {GateType::S, &ErrorAnalyzer::H_XY},
-          {GateType::S_DAG, &ErrorAnalyzer::H_XY},
-          {GateType::SQRT_XX, &ErrorAnalyzer::SQRT_XX},
-          {GateType::SQRT_XX_DAG, &ErrorAnalyzer::SQRT_XX},
-          {GateType::SQRT_YY, &ErrorAnalyzer::SQRT_YY},
-          {GateType::SQRT_YY_DAG, &ErrorAnalyzer::SQRT_YY},
-          {GateType::SQRT_ZZ, &ErrorAnalyzer::SQRT_ZZ},
-          {GateType::SQRT_ZZ_DAG, &ErrorAnalyzer::SQRT_ZZ},
-          {GateType::SWAP, &ErrorAnalyzer::SWAP},
-          {GateType::ISWAP, &ErrorAnalyzer::ISWAP},
-          {GateType::ISWAP_DAG, &ErrorAnalyzer::ISWAP},
-          {GateType::CXSWAP, &ErrorAnalyzer::CXSWAP},
-          {GateType::SWAPCX, &ErrorAnalyzer::SWAPCX},
-      })) {
+      gate_vtable(error_analyzer_vtable_data()) {
 }
 
 void ErrorAnalyzer::run_circuit(const Circuit &circuit) {
